Rename random to random_opt, move builder to utils/
