+++ simulation/common/__init__.py
@@ -3,4 +3,3 @@ from simulation.common.exceptions import BatteryEmptyError
 from simulation.common.helpers import timeit
 from simulation.common.producer import Producer
 from simulation.common.storage import Storage
-from simulation.common.simulationBuilder import SimulationBuilder
