+++ simulation/run_simulation.py
@@ -3,13 +3,12 @@ import datetime
 import numpy as np
 import json
 import sys
-
-from main.Simulation import SimulationReturnType
+from main.Simulation import Simulation, SimulationReturnType
 from optimization.bayesian import BayesianOptimization
-from optimization.random import RandomOptimization
+from optimization.random_opt import RandomOptimization
 from utils.InputBounds import InputBounds
 from config import config_directory
-from common.simulationBuilder import SimulationBuilder
+from utils.SimulationBuilder import SimulationBuilder
 
 """
 Description: Execute simulation optimization sequence.
