+++ simulation/utils/__init__.py
@@ -1,1 +1,2 @@
 from simulation.utils.InputBounds import InputBounds
+from simulation.utils.SimulationBuilder import SimulationBuilder
